backend/requirements.txt
Flask==3.0.2
flask-cors==4.0.0
gunicorn==22.0.0
psycopg2-binary==2.9.10
numpy==2.2.3
pandas==2.2.3
yfinance==0.2.40
peewee==3.17.6
requests==2.32.3
python-dateutil==2.9.0.post0
